+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Created Campus Guide home widget [#1808](https://github.com/rokwire/illinois-app/issues/1808).
+- Created Dining home widget [#1814](https://github.com/rokwire/illinois-app/issues/1814).
 
 ## [4.0.35] - 2022-06-22
 ### Fixed